safari-600.7-branch/LayoutTests/http/tests/w3c/webperf/submission/Google/resource-timing/html/test_resource_reparenting.html

<!DOCTYPE html>
<html>
<head>
<meta charset="utf-8" />
<title>Resource Timing reparenting elements</title>
<link rel="author" title="Google" href="http://www.google.com/" />
<link rel="help" href="http://www.w3.org/TR/resource-timing/#dom-performanceresourcetiming-initiatortype"/>
<script src="/w3c/resources/testharness.js"></script>
<script src="/w3c/resources/testharnessreport.js"></script>
<script src="/w3c/webperf/resources/webperftestharness.js"></script>
<script src="/w3c/webperf/resources/webperftestharnessextension.js"></script>
<script>
var iframe;
function setup_iframe() {
    iframe = document.getElementById('frameContext');
    var d = iframe.contentWindow.document;
    var iframeBody = d.createElement('body');

    var move_to_parent = d.createElement("img");
    move_to_parent.src = '/w3c/webperf/resources/generate_resource.php?type=image&id=move_to_parent';
    iframeBody.appendChild(move_to_parent);
    iframeBody.removeChild(move_to_parent);

    var parentBody = document.getElementsByTagName('body')[0];
    parentBody.appendChild(move_to_parent);

    var move_to_child = document.createElement("img");
    move_to_child.src = '/w3c/webperf/resources/generate_resource.php?type=image&id=move_to_child';
    parentBody.appendChild(move_to_child);
    parentBody.removeChild(move_to_child);
    iframeBody.appendChild(move_to_child);
}
function onload_test() {
    var context = new PerformanceContext(iframe.contentWindow.performance);
    var entries = context.getEntries();

    var expected_entries = {
        '/w3c/webperf/resources/generate_resource.php?type=image&id=move_to_parent': 'img',
    };

    test_resource_entries(entries, expected_entries);
}
window.setup_iframe = setup_iframe;
</script>
</head>
<body>
<h1>Description</h1>
<p>This test validates that reparenting an element doesn't change the initiator document.</p>
<div id="log"></div>
<iframe id="frameContext" onload="onload_test();" src="/w3c/webperf/resources/inject_resource_test.html"></iframe>
</body>
</html>
