LayoutTests/http/tests/w3c/webperf/submission/Google/resource-timing/html/test_resource_reparenting-expected.txt

Description

This test validates that reparenting an element doesn't change the initiator document.

  

PASS /w3c/webperf/resources/generate_resource.php?type=image&id=move_to_parent is expected to have initiatorType img 
